requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
sqlparse==0.4.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2